% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_traits}
\alias{add_traits}
\title{Adding diet category and stable isotopes}
\usage{
add_traits(result_ae, sia_species, intestine)
}
\arguments{
\item{result_ae}{}

\item{sia_species}{sia means on species level}
}
\value{

}
\description{
Adding diet category and stable isotopes
}
