% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ae_mod}
\alias{ae_mod}
\title{Run model to get assimilation efficiency}
\usage{
ae_mod(model, data, element)
}
\arguments{
\item{model}{The compiled stan model}

\item{data}{The data containing cnp\%}

\item{element}{The element (C, N or P)}
}
\value{

}
\description{
Run model to get assimilation efficiency
}
