% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_ae_models}
\alias{run_ae_models}
\title{Run ae models for all species}
\usage{
run_ae_models(data, stanmodel)
}
\arguments{
\item{data}{}

\item{stanmodel}{}
}
\value{

}
\description{
Run ae models for all species
}
