% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plots.R
\name{save_plot}
\alias{save_plot}
\title{Title}
\usage{
save_plot(plot, name, width = 8, height = 6)
}
\arguments{
\item{plot}{}

\item{name}{}

\item{width}{}

\item{height}{}
}
\value{

}
\description{
Title
}
