% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_traits}
\alias{add_traits}
\title{Adding diet category and stable isotopes}
\usage{
add_traits(result_ae)
}
\arguments{
\item{result_ae}{}
}
\value{

}
\description{
Adding diet category and stable isotopes
}
